+++ src/Fpl.SearchConsole/SimpleLeagueIndexBookmarkProvider.cs
@@ -1,13 +1,20 @@
 using System;
 using System.Threading.Tasks;
 using Fpl.Search.Indexing;
+using Microsoft.Extensions.Logging;
 
 namespace Fpl.SearchConsole
 {
     internal class SimpleLeagueIndexBookmarkProvider : IIndexBookmarkProvider
     {
+        private readonly ILogger<SimpleLeagueIndexBookmarkProvider> _logger;
         private string Path = "./bookmark.txt";
 
+        public SimpleLeagueIndexBookmarkProvider(ILogger<SimpleLeagueIndexBookmarkProvider> logger)
+        {
+            _logger = logger;
+        }
+        
         public async Task<int> GetBookmark()
         {
             try
@@ -18,7 +25,7 @@ namespace Fpl.SearchConsole
             }
             catch (Exception e)
             {
-                Console.WriteLine(e);
+                _logger.LogError(e, e.Message);
                 return 1;
             }
         }
@@ -27,12 +34,12 @@ namespace Fpl.SearchConsole
         {
             try
             {
-                Console.WriteLine($"Setting bookmark at {bookmark}.");
+                _logger.LogInformation($"Setting bookmark at {bookmark}.");
                 return System.IO.File.WriteAllTextAsync(Path, bookmark.ToString());
             }
             catch (Exception e)
             {
-                Console.WriteLine(e);
+                _logger.LogError(e, e.Message);
                 return Task.CompletedTask;
             }
         }
